+++ lib/src/flutter_emoji_feedback.dart
@@ -3,36 +3,50 @@ import 'package:flutter/material.dart';
 import 'package:flutter_emoji_feedback/flutter_emoji_feedback.dart';
 import 'package:flutter_svg/flutter_svg.dart';
 
-typedef PresetBuilder = Widget Function(int, EmojiModel);
+typedef EmojiBuilder = Widget Function(int, EmojiModel);
 
 class EmojiFeedback extends StatefulWidget {
   EmojiFeedback({
     Key? key,
     this.onChanged,
-    this.preset = classicEmojiPreset,
+    this.emojiPreset = classicEmojiPreset,
     this.presetBuilder,
     this.showLabel = true,
     this.labelTextStyle,
     this.animDuration = const Duration(milliseconds: 150),
-  })  : assert(preset.isNotEmpty),
+    this.customLabels,
+    this.inactiveElementBlendColor,
+    this.inactiveElementScale,
+    this.elementSize,
+    this.curve,
+  })  : assert(emojiPreset.isNotEmpty),
+        assert(
+          () {
+            if (customLabels != null) {
+              return customLabels.length == emojiPreset.length;
+            }
+            return true;
+          }(),
+          'Please, emojiPreset and customLabels should have the same length',
+        ),
         super(key: key);
 
   /// Function called when an item is selected.
-  /// Values goes from 1 to `preset.length` 
-  /// 
+  /// Values goes from 1 to `preset.length`
+  ///
   /// For the classicEmojiPresset, which length is 5, the values will go from 0 to 5
   final ValueChanged<int>? onChanged;
 
   /// List of emojis
   /// Defaults to `classicEmojiPreset`
-  final List<EmojiModel> preset;
+  final List<EmojiModel> emojiPreset;
 
   /// Duration of the scale animation
   /// Defaults to `const Duration(milliseconds: 150)`
   final Duration animDuration;
 
   /// Custom emoji widget builder
-  final PresetBuilder? presetBuilder;
+  final EmojiBuilder? presetBuilder;
 
   /// Wether the label should be displayed or not
   /// Defaults to `true`
@@ -40,6 +54,26 @@ class EmojiFeedback extends StatefulWidget {
 
   /// Style of the emoji label
   final TextStyle? labelTextStyle;
+
+  /// Define your custom labels using this attribute.
+  /// Useful if you wish to use the predefined emojis with custom labels
+  final List<String>? customLabels;
+
+  /// Custom blend color for inactive elements
+  /// Defaults to `Colors.grey`
+  final Color? inactiveElementBlendColor;
+
+  /// Scale factor of inactiveElements
+  /// Defaults to `0.7`
+  final double? inactiveElementScale;
+
+  /// Size of emoji elements
+  /// Default value is calculated so each element occupate the same amount of space
+  /// Formula: `(maxWidth / emojiCount) - spaceBetweenItems`
+  final double? elementSize;
+
+  /// Scale animation curve
+  final Curve? curve;
 
   @override
   State<EmojiFeedback> createState() => _EmojiFeedbackState();
@@ -60,14 +94,15 @@ class _EmojiFeedbackState extends State<EmojiFeedback> {
     return LayoutBuilder(
       builder: (context, constraints) {
         const double spaceBetween = 10.0;
-        const double inactiveElementScale = .7;
-        final double elementSize = (constraints.maxWidth / widget.preset.length) - spaceBetween;
+        final double inactiveElementScale = widget.inactiveElementScale ?? .7;
+        final double elementSize =
+            widget.elementSize ?? (constraints.maxWidth / widget.emojiPreset.length) - spaceBetween;
 
         return Row(
           crossAxisAlignment: CrossAxisAlignment.end,
           mainAxisAlignment: MainAxisAlignment.spaceBetween,
           children: List.from(
-            widget.preset.mapIndexed(widget.presetBuilder ??
+            widget.emojiPreset.mapIndexed(widget.presetBuilder ??
                 (index, element) {
                   final isItemActive = activeItemIndex == index;
 
@@ -75,12 +110,12 @@ class _EmojiFeedbackState extends State<EmojiFeedback> {
                     element.src,
                     width: elementSize,
                     package: element.package,
-                    colorBlendMode: isItemActive ? BlendMode.saturation : BlendMode.srcIn,
                   );
 
                   return AnimatedScale(
                     scale: isItemActive ? 1 : inactiveElementScale,
                     duration: widget.animDuration,
+                    curve: widget.curve ?? Curves.linear,
                     child: Column(
                       children: [
                         isItemActive
@@ -90,8 +125,8 @@ class _EmojiFeedbackState extends State<EmojiFeedback> {
                                   setActiveItem(index);
                                 },
                                 child: Container(
-                                  foregroundDecoration: const BoxDecoration(
-                                    color: Colors.grey,
+                                  foregroundDecoration: BoxDecoration(
+                                    color: widget.inactiveElementBlendColor ?? Colors.grey,
                                     backgroundBlendMode: BlendMode.saturation,
                                   ),
                                   child: child,
@@ -99,7 +134,10 @@ class _EmojiFeedbackState extends State<EmojiFeedback> {
                               ),
                         if (widget.showLabel) ...[
                           const SizedBox(height: 5.0),
-                          Text(element.label, style: widget.labelTextStyle),
+                          Text(
+                            widget.customLabels?.elementAt(index) ?? element.label,
+                            style: widget.labelTextStyle,
+                          ),
                         ]
                       ],
                     ),